refactor: split oversized editor.ts and document-routes.ts under 200-line limit
Extract collab/awareness/event-listener wiring from editor.ts into editor-collab.ts,
and extract per-document read/mutation/delete routes from document-routes.ts into
document-mutation-routes.ts, keeping both originals under the 200-line ceiling.
